+++ src/LanguageServer/Protocol/Features/Suggestions/UnifiedPredefinedSuggestedActionCategoryNames.cs
@@ -2,7 +2,7 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-namespace Microsoft.CodeAnalysis.UnifiedSuggestions;
+namespace Microsoft.CodeAnalysis.Suggestions;
 
 /// <summary>
 /// Equivalent to PredefinedSuggestedActionCategoryNames, but in a location that
