+++ src/LanguageServer/Protocol/Features/Suggestions/UnifiedSuggestedActionSetComparer.cs
@@ -8,7 +8,7 @@ using System;
 using System.Collections.Generic;
 using Microsoft.CodeAnalysis.Text;
 
-namespace Microsoft.CodeAnalysis.UnifiedSuggestions;
+namespace Microsoft.CodeAnalysis.Suggestions;
 
 internal sealed class UnifiedSuggestedActionSetComparer : IComparer<UnifiedSuggestedActionSet>
 {
